move service-card.tsx
